+++ api/src/interfaces/propsModel.ts
@@ -1,18 +1,3 @@
-export interface userProps {
-	id: number;
-	name: string;
-	lastName: string;
-	password: string;
-	email: string;
-	image: string;
-}
-
-export interface loginData {
-	email: string
-	password: string | number
-}
-
-
 export interface productProps {
 	id: number;
 	name: string;
